rename test_harness_aten -> test_harness_aten_cuda
ATen library can be build without CUDA.  test_harness_aten additionally
includes and uses CUDA library.  Make it obvious from its name.
