+++ examples/example_tmm.cc
@@ -27,7 +27,7 @@
 #include "tc/core/mapping_options.h"
 
 #include "../test/test_harness.h"
-#include "../test/test_harness_aten.h"
+#include "../test/test_harness_aten_cuda.h"
 #include "example_fixture.h"
 
 #include "tc/c2/context.h"
