+++ test/test_execution_engine.cc
@@ -27,7 +27,7 @@
 #include "tc/core/mapping_options.h"
 #include "tc/library/common.h"
 
-#include "test_harness_aten.h"
+#include "test_harness_aten_cuda.h"
 
 struct ATenCompilationUnitTest : public ::testing::Test {
   static constexpr uint32_t N = 8, C = 16, O = 6, H = 24, W = 27;
