+++ test/test_tc_mapper.cc
@@ -24,7 +24,7 @@
 #include "tc/core/cuda.h"
 #include "tc/core/scope_guard.h"
 
-#include "test_harness_aten.h"
+#include "test_harness_aten_cuda.h"
 
 using namespace std;
